An Ant Task for Cruise Control .NET
A Cruise Control .NET Ant Task.  Currently CCNet supports NAnt but not
Ant natively, this task adds support for Ant.  It is based solely on the
NAntTask class from the CCNet Core project, just modified to output
parameters in the form that Ant requires them and defaulting to
different options.